Add README for testing, add init, and small refactor
