+++ echopype/tests/README.md
@@ -1,0 +1,30 @@
+# Echopype tests
+
+This directory contains all the various tests for echopype.
+
+Sample datasets for testing can be found one directory above this one called `test_data`.
+The actual datasets are not stored within the repository, but are stored in a Google Drive folder.
+When the testing infrastructures are spun up, data are pulled from the Google Drive into the repository to be used for testing.
+
+## Current structure
+
+Currently the tests are separated by modules, mirroring a similar structure of the software modules.
+Within each test modules, a mix of *integration* and *unit* testings can be found.
+The following modules are currently being tested:
+
+- calibrate
+- consolidate
+- convert
+- echodata
+- metrics
+- preprocess
+- utils
+- visualize
+- core
+
+Throughout each module, there are special file called `conftest.py`.
+This file contains all of the fixtures used within that module.
+However, the `conftest.py` within the root of the `tests` directory should contain fixtures that can be used by all tests,
+therefore, the fixtures `scope` must be `"session"`.
+
+Within echopype's software module there's a module called `testing`. Within this module contains any helper functions or classes to be used in testing.
