+++ echopype/tests/echodata/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+from echopype.testing import get_mock_echodata
+
+
+@pytest.fixture
+def mock_echodata():
+    return get_mock_echodata()
